Version 1 of mod

+++ preload/shaders/InverseDots.frag
@@ -6,11 +6,11 @@
 uniform float _amount;
 
 vec4 dots(vec4 color) {
-  float xPix = floor(openfl_TextureCoordv.x * openfl_TextureSize.x );
-  float yPix = floor(openfl_TextureCoordv.y * openfl_TextureSize.y);
+  float xPix = float(floor(openfl_TextureCoordv.x * openfl_TextureSize.x));
+  float yPix = float(floor(openfl_TextureCoordv.y * openfl_TextureSize.y));
 
-  int x = int(mod(xPix, 6.0));
-  int y = int(mod(yPix, 4.0));
+  float x = float(int(mod(xPix, 6.0)));
+  float y = float(int(mod(yPix, 4.0)));
 
   if (color.a > 0.0 && ((x == 0.0 && y == 0.0) || (x == 3.0 && y == 2.0)))
   {
